+++ src/server.ts
@@ -3,11 +3,10 @@ import { errorHandler } from './middlewares/errorHandler.middleware';
 
 import express, { Request, Response } from 'express'
 //import { Request, Response, NextFunction ,ErrorRequestHandler} from 'express'
-import cors from 'cors'
 // import morgan from 'morgan'
 import cookieParser from "cookie-parser"
 import {mainRouter} from './router'
-import { APP_BASE_URL,CORS_ORIGINS,NODE_ENV } from './config/config';
+import { APP_BASE_URL,NODE_ENV } from './config/config';
 import { notFoundRouter } from './routes/notFound';
 import morgan from 'morgan'
 import swaggerUI from 'swagger-ui-express'
@@ -18,6 +17,7 @@ import {
   jsonBodyParser,
   urlEncodedBodyParser,
 } from './middlewares/requestBody.middleware';
+import { corsMiddleware } from './middlewares/cors.middleware';
 
 type ServerDependencies = {
   checkDatabase: () => Promise<void>;
@@ -62,29 +62,7 @@ export const createServer = async (
     //Notre serveur parsera les requête entrante en Json
     // server.use(express.json()) 
     server.use(cookieParser());
-    //On indique les cors (qui peut emettre des call depuis notre API)
-    // blocking cors errors:
-  let origin:string | Array<string>=""
-  // console.log(NODE_ENV)
-  if (NODE_ENV=='production'){
-    origin="https://myhappywallet.andriacapai.com"
-  }
-  if (NODE_ENV=='development'){
-    origin=CORS_ORIGINS.length > 0
-      ? CORS_ORIGINS
-      : ["http://localhost:3000","http://localhost:5173","https://myhappywallet.andriacapai.com"]
-  }
-  
-  const corsOptions = {
-    origin: origin,
-    credentials: true,           
-    methods: ["OPTIONS,GET,HEAD,PUT,PATCH,POST,DELETE"],
-    // "preflightContinue": true,
-    optionSuccessStatus: 200,
-  }
-    server.use(cors(
-      corsOptions
-    ))
+    server.use(corsMiddleware)